requirements.txt
numpy
numpy-stl
trimesh
pyglet
python-utils 